+++ packages/query/CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- WS libarary with support for graphiql (#2600)
+- Subscription type from JSON to the correct entity type (#2600)
 
 ## [2.16.0] - 2024-11-22
 ### Changed